# backend/src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "API de Mensageria",
    "description": "API para envio de mensagens via Whatsapp"
  },
  "host": "http://localhost:3333",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "x-auth-api",
      "description": "API Key para ter acesso ao endpoint"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/{connectionId}/connect": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "connectionId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "x-api-key",
            "in": "header",
            "type": "string"
          },
          {
            "name": "x-auth-api",
            "in": "header",
            "type": "string"
          },
          {
            "name": "connection",
            "in": "query",
            "type": "string"
          },
          {
            "name": "token",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/{connectionId}/disconnect": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "connectionId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "x-api-key",
            "in": "header",
            "type": "string"
          },
          {
            "name": "x-auth-api",
            "in": "header",
            "type": "string"
          },
          {
            "name": "connection",
            "in": "query",
            "type": "string"
          },
          {
            "name": "token",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/{connectionId}/send": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "connectionId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "x-api-key",
            "in": "header",
            "type": "string"
          },
          {
            "name": "x-auth-api",
            "in": "header",
            "type": "string"
          },
          {
            "name": "connection",
            "in": "query",
            "type": "string"
          },
          {
            "name": "token",
            "in": "query",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "phone": {
                  "example": "any"
                },
                "message": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/sync": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "type": "string"
          },
          {
            "name": "x-auth-api",
            "in": "header",
            "type": "string"
          },
          {
            "name": "connection",
            "in": "query",
            "type": "string"
          },
          {
            "name": "token",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
...
}